add tests for metadata/coosyses
