+++ astropy/io/votable/tests/data/table_from_paper.vot
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<VOTABLE version="1.4" xmlns="http://www.ivoa.net/xml/VOTable/v1.3">
+  <RESOURCE>
+    <COOSYS ID="system" epoch="J2015.5" system="ICRS"/>
+    <TIMESYS ID="time_frame" refposition="BARYCENTER" timeorigin="2455197.5" timescale="TCB"/>
+    <TABLE name="ts_data">
+      <FIELD datatype="double" name="obs_time" ucd="time.epoch" unit="d" ref="time_frame"/>
+      <FIELD datatype="float" name="flux" ucd="phot.flux;em.opt.V" unit="s**-1"/>
+      <FIELD datatype="float" name="mag" ucd="phot.mag;em.opt.V" unit="mag"/>
+      <FIELD datatype="float" name="flux_error" ucd="stat.error;phot.flux;em.opt.V" 
+        unit="s**-1"/>
+      <PARAM datatype="double" name="ra" ucd="pos.eq.ra" value="45.7164887146879" ref="system"/>
+      <PARAM datatype="double" name="dec" ucd="pos.eq.dec" value="1.18583048057467" ref="system"/>
+      <DATA>
+        <TABLEDATA>
+          <TR>
+            <TD>1821.2846388435</TD>
+            <TD>168.358</TD>
+            <TD>20.12281560517953</TD>
+            <TD>8.71437</TD>
+          </TR>
+        </TABLEDATA>
+      </DATA>
+    </TABLE>
+  </RESOURCE>
+</VOTABLE>
